[proliferate] split preference domain helpers
