+++ desktop/src/hooks/workspaces/tabs/use-workspace-shell-activation.test.tsx
@@ -11,7 +11,7 @@ import { useSessionSelectionStore } from "@/stores/sessions/session-selection-st
 import { useSessionTranscriptStore } from "@/stores/sessions/session-transcript-store";
 import {
   WORKSPACE_UI_DEFAULTS,
-} from "@/lib/domain/preferences/workspace-ui-state";
+} from "@/lib/domain/preferences/workspace-ui/model";
 import { useWorkspaceUiStore } from "@/stores/preferences/workspace-ui-store";
 import { useWorkspaceShellActivation } from "@/hooks/workspaces/tabs/use-workspace-shell-activation";
