+++ desktop/src/hooks/workspaces/use-workspace-activity-acknowledgement.test.tsx
@@ -5,7 +5,7 @@ import type { LogicalWorkspace } from "@/lib/domain/workspaces/cloud/logical-wor
 import { makeLocalLogicalWorkspace } from "@/lib/domain/workspaces/sidebar/sidebar-test-fixtures";
 import {
   WORKSPACE_UI_DEFAULTS,
-} from "@/lib/domain/preferences/workspace-ui-state";
+} from "@/lib/domain/preferences/workspace-ui/model";
 import { useWorkspaceUiStore } from "@/stores/preferences/workspace-ui-store";
 import { useSessionSelectionStore } from "@/stores/sessions/session-selection-store";
 import { useWorkspaceActivityAcknowledgement } from "./use-workspace-activity-acknowledgement";
