+++ desktop/src/lib/domain/workspaces/cloud/cloud-workspace-creation.ts
@@ -5,7 +5,7 @@ import {
 } from "@/lib/domain/workspaces/cloud/cloud-workspace-model";
 import type { CloudRepoConfigSummary } from "@/lib/domain/cloud/repo-configs";
 import type { AuthUser } from "@/lib/domain/auth/auth-user";
-import type { BranchPrefixType } from "@/lib/domain/preferences/user-preferences";
+import type { BranchPrefixType } from "@/lib/domain/preferences/user/model";
 import { generateWorkspaceSlug } from "@/lib/domain/workspaces/creation/arrival";
 import {
   buildBranchName,
